test/htmlwg/submission/Ms2ger/events/event-handler-spec-example.html

<!DOCTYPE html>
<title>Event handler invocation order</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<button type="button" id="test">Start test</button>
<script>
var t = async_test("Event handler listeners should be registered when they are first set.");
var i = 0;
var uncalled = "t.step(function() { assert_unreached('First event handler.') })"
var button = document.getElementById('test');
button.onclick = {};

// XXX:
// The spec has changed in how to deal with handlers registered
// on idl or content attributes rather than with addEventListener.
// Used to say that they always ran first.  That's how webkit does it
// Spec now says that they run in a position determined by
// when they were first set.  FF does it that way.  See:
// http://www.whatwg.org/specs/web-apps/current-work/multipage/webappapis.html#event-handlers

button.addEventListener('click', t.step_func(function () { assert_equals(++i, 2) }), false);
button.setAttribute('onclick', uncalled); // event handler listener is registered here
button.addEventListener('click', t.step_func(function () { assert_equals(++i, 3) }), false);
button.onclick = t.step_func(function () { assert_equals(++i, 1); });
button.addEventListener('click', t.step_func(function () { assert_equals(++i, 4) }), false);
button.click()
assert_equals(button.getAttribute("onclick"), uncalled)
assert_equals(i, 4);
t.done()
</script>
